출석 체크 시 유효성 검사, QRcode 테이블 등록 및 자동 삭제

+++ gym/src/main/java/com/gym/gym/domain/Attendance.java
@@ -11,7 +11,7 @@ import lombok.Data;
 public class Attendance {
 
     private String qrId; // QR 코드 ID
-    private int userNo; // 회원 번호
+    private Long userNo; // 회원 번호
     private Date checkTime; // 출석 시간
     private Users users; // Users 객체 (회원 정보)
     private int attendanceCount; // 출석 횟수
